[x86/x64] Add support for x86-64 and improve 32 bit mode too.
This adds support for x86-64 dynarec both on Windows and Linux. Since
they have different requirements there's some macro magic in the stubs
file.

This also fixes x86 support in some cases: stack alignment requirements
where violated all over. This allows the usage of clang as a compiler
(which has a tendency to use SSE instructions more often than gcc does).

To support this I also reworked the mmap/VirtualAlloc magic to make sure
JIT arena stays close to .text.

Fixed some other minor issues and removed some unnecessary JIT code here
and there. clang tends to do some (wrong?) assumptions about global
symbols alignment.

+++ cpu_threaded.c
@@ -2937,7 +2937,7 @@ block_data_type block_data[MAX_BLOCK_SIZE];
 block_exit_type block_exits[MAX_EXITS];
 
 #define smc_write_arm_yes() {                                                 \
-  int offset = (pc < 0x03000000) ? 0x40000 : -0x8000;                         \
+  intptr_t offset = (pc < 0x03000000) ? 0x40000 : -0x8000;                    \
   if(address32(pc_address_block, (block_end_pc & 0x7FFF) + offset) == 0)      \
   {                                                                           \
     address32(pc_address_block, (block_end_pc & 0x7FFF) + offset) =           \
@@ -2946,7 +2946,7 @@ block_exit_type block_exits[MAX_EXITS];
 }
 
 #define smc_write_thumb_yes() {                                               \
-  int offset = (pc < 0x03000000) ? 0x40000 : -0x8000;                         \
+  intptr_t offset = (pc < 0x03000000) ? 0x40000 : -0x8000;                    \
   if(address16(pc_address_block, (block_end_pc & 0x7FFF) + offset) == 0)      \
   {                                                                           \
     address16(pc_address_block, (block_end_pc & 0x7FFF) + offset) =           \
